Move function definitions to separate module + fix linting errors
