+++ utils/FitUtils.py
@@ -1,0 +1,121 @@
+'''
+Module with function definitions and fit utils
+'''
+
+import numpy as np
+from ROOT import TMath # pylint: disable=import-error,no-name-in-module
+
+def SingleGaus(x, par):
+    '''
+    Gaussian function
+
+    Parameters
+    ----------
+
+    - x: function variable
+    - par: function parameters
+        par[0]: normalisation
+        par[1]: mean
+        par[2]: sigma
+    '''
+    return par[0]*TMath.Gaus(x[0], par[1], par[2], True)
+
+
+def DoubleGaus(x, par):
+    '''
+    Sum of two Gaussian functions with same mean and different sigma
+
+    Parameters
+    ----------
+
+    - x: function variable
+    - par: function parameters
+        par[0]: normalisation
+        par[1]: mean
+        par[2]: first sigma
+        par[3]: second sigma
+        par[4]: fraction of integral in second Gaussian
+    '''
+    firstGaus = TMath.Gaus(x[0], par[1], par[2], True)
+    secondGaus = TMath.Gaus(x[0], par[1], par[3], True)
+    return par[0] * ((1-par[4])*firstGaus + par[4]*secondGaus)
+
+
+def DoublePeakSingleGaus(x, par):
+    '''
+    Sum of two Gaussian functions with different mean and sigma
+
+    Parameters
+    ----------
+
+    - x: function variable
+    - par: function parameters
+        par[0]: normalisation first peak
+        par[1]: mean first peak
+        par[2]: sigma first peak
+        par[3]: normalisation second peak
+        par[4]: mean second peak
+        par[5]: sigma second peak
+    '''
+    firstGaus = par[0]*TMath.Gaus(x[0], par[1], par[2], True)
+    secondGaus = par[3]*TMath.Gaus(x[0], par[4], par[5], True)
+    return firstGaus + secondGaus
+
+
+def DoublePeakDoubleGaus(x, par):
+    '''
+    Sum of a double Gaussian function and a single Gaussian function
+
+    Parameters
+    ----------
+
+    - x: function variable
+    - par: function parameters
+        par[0]: normalisation first peak
+        par[1]: mean first peak
+        par[2]: first sigma first peak
+        par[3]: second sigma first peak
+        par[4]: fraction of integral in second Gaussian first peak
+        par[5]: normalisation second peak
+        par[6]: mean second peak
+        par[7]: sigma second peak
+    '''
+    firstGaus = TMath.Gaus(x[0], par[1], par[2], True)
+    secondGaus = TMath.Gaus(x[0], par[1], par[3], True)
+    thirdGaus = par[5]*TMath.Gaus(x[0], par[6], par[7], True)
+    return par[0] * ((1-par[4])*firstGaus + par[4]*secondGaus) + thirdGaus
+
+
+def VoigtFunc(x, par):
+    '''
+    Voigtian function
+
+    Parameters
+    ----------
+
+    - x: function variable
+    - par: function parameters
+        par[0]: normalisation
+        par[1]: mean
+        par[2]: sigma
+        par[3]: gamma
+    '''
+
+    return par[0] * TMath.Voigt(x[0]-par[1], par[2], par[3])
+
+
+def ExpoPowLaw(x, par):
+    '''
+    Exponential times power law function
+
+    Parameters
+    ----------
+
+    - x: function variable
+    - par: function parameters
+        par[0]: normalisation
+        par[1]: mass (lowest possible value)
+        par[2]: expo slope
+    '''
+
+    return par[0] * np.sqrt(x[0] - par[1]) * np.exp(-1. * par[2] * (x[0] - par[1]))
